% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut-expand_grid_df.R
\name{expand_grid_df}
\alias{expand_grid_df}
\title{Expand on two data.frames}
\usage{
expand_grid_df(data_1, data_2)
}
\arguments{
\item{data_1}{first data.frame}

\item{data_2}{second data.frame}
}
\description{
An extension of expand.grid() to expand on two \code{data.frame}s
}
