% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-st_intesects_join.R
\name{st_intersects_join}
\alias{st_intersects_join}
\title{st_intersects_join}
\usage{
st_intersects_join(sf_1, sf_2)
}
\arguments{
\item{sf_1}{An sf object}

\item{sf_2}{An sf object}
}
\description{
This function joins the attributes of two sf objects based on st_intesects()
with geometry information removed. This is much faster than st_intersection().
}
