% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-st_shift.R
\name{st_shift}
\alias{st_shift}
\title{st_shift}
\usage{
st_shift(data_sf, shift, keep_data = TRUE)
}
\arguments{
\item{data_sf}{An sf object}

\item{shift}{A vector of two numeric numbers (x, y): data_sf is shifted by (x, y)}

\item{keep_data}{keep the non-geometry data part of the data (default: TRUE)}
}
\description{
shift an sf object
}
