% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sp-st_tilt.R
\name{st_tilt}
\alias{st_tilt}
\title{st_tilt}
\usage{
st_tilt(data_sf, angle, base_sf = NULL, keep_data = TRUE)
}
\arguments{
\item{data_sf}{An sf object}

\item{angle}{(angle to tilt, numeric)}

\item{base_sf}{If supplied, \code{data_sf} is shifted to the centroid of this sf object (default: NULL)}

\item{keep_data}{keep the non-geometry data part of the data (default: TRUE)}
}
\description{
Tilt an sf object by a specified angle
}
